requirements.txt
llama-index
llama-index-vector-stores-qdrant
qdrant-client
pymupdf
python-dotenv
streamlit
ragas
pandas
numpy